fix: 구조체 ID 타입 변경
- uint32 -> int64
- uint와 int 간의 메모리 구조가 다르며, gorm의 쿼리에 집계 함수의 경우 int64로 반환 되기에 통일성을 맞추기 위하여 변경

+++ internal/member/dto.go
@@ -1,7 +1,7 @@
 package member
 
 type GetProfileResponse struct {
-	ID          uint32 `json:"id"`
+	ID          int64  `json:"id"`
 	Name        string `json:"name"`
 	Email       string `json:"email"`
 	PhoneNumber string `json:"phoneNumber,omitempty"`
