+++ internal/model/member.go
@@ -4,7 +4,7 @@ package model
 // Oracle sequence MEMBER_SEQ is used for ID generation
 type Member struct {
 	// Primary key - Oracle IDENTITY (auto-increment)
-	ID uint32 `gorm:"column:id;primaryKey;autoIncrement"`
+	ID int64 `gorm:"column:id;primaryKey;autoIncrement"`
 
 	// Core fields
 	Email       string `gorm:"column:email;type:VARCHAR2(255);not null;uniqueIndex:idx_member_email"` // 이메일 (unique)
