+++ internal/model/member_room.go
@@ -5,11 +5,11 @@ package model
 // Oracle IDENTITY (auto-increment) is used for ID generation
 type MemberRoom struct {
 	// Primary key - Oracle IDENTITY (auto-increment)
-	ID uint32 `gorm:"column:id;primaryKey;autoIncrement"`
+	ID int64 `gorm:"column:id;primaryKey;autoIncrement"`
 
 	// Foreign keys
-	MemberID uint32 `gorm:"column:member_id;not null;uniqueIndex:idx_member_room_member_id_room_id"` // FK to member
-	RoomID   uint32 `gorm:"column:room_id;not null;uniqueIndex:idx_member_room_member_id_room_id"`   // FK to room
+	MemberID int64 `gorm:"column:member_id;not null;uniqueIndex:idx_member_room_member_id_room_id"` // FK to member
+	RoomID   int64 `gorm:"column:room_id;not null;uniqueIndex:idx_member_room_member_id_room_id"`   // FK to room
 
 	// Foreign key relations (Lazy loading in Java)
 	Member *Member `gorm:"foreignKey:MemberID;references:ID;constraint:OnDelete:CASCADE"` // Member 엔티티 참조
@@ -28,7 +28,7 @@ func (*MemberRoom) TableName() string {
 }
 
 // NewRoomMember creates a new MemberRoom instance
-func NewRoomMember(memberID, roomID uint32, role string, isNotification bool) *MemberRoom {
+func NewRoomMember(memberID, roomID int64, role string, isNotification bool) *MemberRoom {
 	return &MemberRoom{
 		MemberID:       memberID,
 		RoomID:         roomID,
